# Updated mempool debug log