Optimized the Mempool loop log output.

Before:

```
Aug 19 17:56:22 [6183] DEBUG: Updating mempool
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 1
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 2
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 3
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 4
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 5
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 6
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 7
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 8
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 9
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 10
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 11
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 12
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 13
Aug 19 17:56:23 [6183] DEBUG: Fetched transaction 14
Aug 19 17:56:23 [6183] DEBUG: Mempool blocks calculated in 0.007 seconds
Aug 19 17:56:23 [6183] DEBUG: New mempool size: 8198 Change: -251
Aug 19 17:56:23 [6183] DEBUG: Mempool updated in 0.129 seconds
```

After:

```
Aug 19 19:35:44 [7261] DEBUG: Updating mempool...
Aug 19 19:35:44 [7261] DEBUG: Mempool blocks calculated in 0.004 seconds
Aug 19 19:35:44 [7261] DEBUG: Mempool updated in 0.121 seconds. New size: 2740 (+13)
```