@wiz 2022-08-20
Thanks for the PR but maybe the correct fix is to disable the mempool loop in the Lightning backend?