@softsimon 2022-08-20
> Thanks for the PR but maybe the correct fix is to disable the mempool loop in the Lightning backend?

As long as the mempool loop is working it should debug some output.
If you want to disable the mempool updating completely (not just silent the log) it should be a separate PR