@wiz 2022-08-21
Yes, that's what I'm saying, instead of disabling the "Fetching transactions" debug log, isn't what we really want to correctly fix the issue of the mempool looper running in the Lightning backend? This way we still get the debug output on the mempool backends but it won't spam our logs on Lightning AND save CPU cycles and load on bitcoind RPC